+++ src/main/java/http/ContentType.java
@@ -1,4 +1,4 @@
-package utils;
+package http;
 
 import java.util.Arrays;
 import java.util.Objects;
