+++ src/main/java/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package utils;
+package http;
 
 public enum HttpStatus {
     OK(200, "OK"),
