+++ src/main/java/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package utils;
+package http;
 
 public enum HttpVersion {
