+++ src/main/java/utils/ContentTypeParser.java
@@ -1,5 +1,7 @@
 package utils;
 
+import http.ContentType;
+
 import java.net.FileNameMap;
 import java.net.URLConnection;
