+++ src/main/java/utils/ResponseWriter.java
@@ -1,11 +1,14 @@
 package utils;
 
+import http.Header;
+import http.HttpResponse;
+import http.HttpStatus;
+import http.HttpVersion;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.util.Arrays;
 
 
 public class ResponseWriter {
