+++ src/test/java/utils/ContentTypeParserTest.java
@@ -1,5 +1,6 @@
 package utils;
 
+import http.ContentType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
