+++ src/test/java/utils/HttpMethodTest.java
@@ -1,5 +1,6 @@
 package utils;
 
+import http.HttpMethod;
 import org.junit.jupiter.api.Test;
 
 import java.util.NoSuchElementException;
